Remove unneeded NULL check for type reference.
Compilation with GCC 8 fails due to -Werror=nonnull-compare, where
the compiler assumes that the value will never be NULL, like this case.
The type SFmAlarmDataT &data is a reference (not a pointer), so the
check for NULL is not needed.

Also, a caller to this function was updated accordingly.

Story: 2003498
Task: 24775

Change-Id: Ica07ec8d7ad009e1048f014b0a6440c76de09eef
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>