Change fm_snmp_util_gen_trap to receive a pointer instead of reference
As C++ references cannot be NULL, new compilers optimizes the code
assuming this condition. Therefore, in new compilers, like GCC 8 the
fm_snmp_util_gen_trap will segfault trying to deference
data.entity_instance_id as the if (&data != NULL) will be always
optimized to be true.

This behavior has been tested on GCC 9, 8, 7 and 4, where the newer
versions (9, 8 and 7) shows this behavior. The proposed solution is to
change the signature of fm_snmp_util_gen_trap to receive a pointer
instead of a reference. Other functions in this project, like fm_set_uuid
or fm_tr_alarm_id receives SFmAlarmDataT * as argument and there is no
evident reasons to use references in this function.

The signature of fm_snmp_util_gen_trap was changed and the callers updated
accordingly.

Story: 2003498
Task: 24775

Change-Id: Ica07ec8d7ad009e1048f014b0a6440c76de09eef
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>