Make fm-common build a bit more portable
Specifically get the psql include path from pg_config as libpq-fe.h
is in different places on RHEL-based and Debian-based systems.

Also add -f to symlink command in install because we can't assume
DEST_DIR is empty.

This is in support of Austin's DevStack wirk in stx-config started
in https://review.openstack.org/#/c/586443/, in my case because all of
my DevStack VMs run Ubuntu xenial.

Needed-by: https://review.openstack.org/595865
Change-Id: I8db0579225487e6b31fe8a5ed88a29820bda7d69
Signed-off-by: Dean Troyer <dtroyer@gmail.com>