Add DevStack support for FM dependencies
Add a DevStack plugin to build and install the basic Fault dependencies
required by most of the other STX projects.  The bulk of this was originally
submitted in https://review.openstack.org/#/c/595865/.

Installing and starting the REST API service will follow in a later review.

Also enable bashate on the devstack plugin and max yaml line length
set to 260

Change-Id: I95009fceee28a81e6d8795e90b259e0e1b175327
Signed-off-by: Dean Troyer <dtroyer@gmail.com>