stx-fault: API ref doc content added.
I added the api-ref-fm-v1-fault.rst file, which is content for
the API reference manual to the api-ref/source dir. This
represents the converted old-style files to the newer OpenStack
supported RST files.

I updated the index.rst file to include the new .rst file so
that the api-ref document can build.

Change-Id: Ie0daae5b4ddc6c6363f647b15c507b664635b2d3
Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>