Fix clear_fault/clear_all in FaultAPIsV2 raise exception when alarm is not found
For the all 6 APIs in FaultAPIsV2, will treat alarm not exist as normal
case, and exception will not be raised. Exception  Exception is raised
only if there is operation failure, such as fmclient fail to connect
fmManager, failure in memory allocation, failure in internal structure
parse, etc.

Test:
Run each 6 APIs from FaultAPIs and FaultAPIsV2 with below case:
1. with correct alarm id
2. with wrong alarm id
3. with fmManager cannot be connected (Modify fmManager listen port to
   simulate it)
Confirm FaultAPIs behavior is the not changed.
And FaultAPIsV2 raises exception for operation failure only.

Closes-Bug: 1821112

Change-Id: I8ba122b19964613d90e9d0bf4a25134ff60e5c19
Signed-off-by: Shuicheng Lin <shuicheng.lin@intel.com>