Add gabbi tests to fm-rest-api functional tests
This adds a basic REST API test framework built around gabbi[0]
that acts as a starting point for REST API tests and validation
that the fm-rest-api service is running and responsive.

It adds the basic setup requirements for running tox tests
in fm-rest-api/fm to the 'functional' tox environment.

OS_TEST_URL points to the faultmanagement endpoint, until we
learn how to get it from the service catalog.

[0] https://github.com/cdent/gabbi

Change-Id: I7f73848fbee4899e70a66e68541712f405f12a44
Signed-off-by: Dean Troyer <dtroyer@gmail.com>