[DevStack] WSME doesn't always get installed, force it
The jobs are working here but stx-nfv devstack fails in fm-rest-api,
what fun it is to find a transient dependency!  fm-rest-api has
apparently never started correctly but we did not properly detect
that until getting the stx-nfv plugin up.

Add requirements.txt and test-requirements.txt, only
test-requirements.txt is directly used by DevStack but WSME
is not a test-only dependency...

Cleanup to follow:
* properly get the dependencies of the service documented for pip
* properly detect when fm-rest-api fails to start under systemd

Change-Id: I91236d2bbb785c75264e8c4772d4ad6dfa45562c
Signed-off-by: Dean Troyer <dtroyer@gmail.com>